Fixed not showing new users in group members table
Fixed an issue which wouldn't show that a user was in the group if there
was no recorded games of them in the games_database. Used a mysql left
join, to include all users, even if they had no records in the 'right'
(games_database) table.

Changed our div element for graph to start of at height, so that content
doesn't 'jump' around when the page is loading